requirements.txt
azure-functions==1.21.3
certifi==2024.12.14
charset-normalizer==3.4.1
duckdb==1.1.3
idna==3.10
requests==2.32.3
urllib3==2.3.0
